requirements.txt
pytest-runner
gino
aiocontextvars
pyyaml
websockets
aiofiles